Add StatisticsController

+++ src/main/java/org/lobox/common/http/filter/requestcounter/RequestCounter.java
@@ -2,18 +2,18 @@ package org.lobox.common.http.filter.requestcounter;
 
 import org.springframework.stereotype.Component;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 @Component
 public class RequestCounter {
 
-    private final AtomicInteger requestCounts = new AtomicInteger(0);
+    private final AtomicLong requestCounts = new AtomicLong(0);
 
     public void countRequest() {
         requestCounts.incrementAndGet();
     }
 
-    public int getRequestCount() {
+    public long getRequestCount() {
         return requestCounts.get();
     }
